feat: `repeat` template processor
