+++ core/template-processors/default-processors.js
@@ -1,0 +1,3 @@
+import { domSetProcessor, refProcessor, repeatProcessor, slotProcessor, txtNodesProcessor } from './index.js';
+
+export const DEFAULT_PROCESSORS = [slotProcessor, refProcessor, repeatProcessor, domSetProcessor, txtNodesProcessor];
