+++ core/template-processors/dom-repeat-set-processor.js
@@ -1,0 +1,4 @@
+import { createDomBindProcessor } from './create-dom-bind-processor.js';
+import { DICT } from '../dictionary.js';
+
+export const domRepeatSetProcessor = createDomBindProcessor(DICT.REPEAT_BIND_ATTR);
