+++ core/template-processors/dom-set-processor.js
@@ -1,0 +1,4 @@
+import { createDomBindProcessor } from './create-dom-bind-processor.js';
+import { DICT } from '../dictionary.js';
+
+export const domSetProcessor = createDomBindProcessor(DICT.BIND_ATTR);
