+++ core/template-processors/index.js
@@ -1,0 +1,5 @@
+export { domSetProcessor } from './dom-set-processor.js';
+export { refProcessor } from './ref-processor.js';
+export { repeatProcessor } from './repeat-processor.js';
+export { slotProcessor } from './slot-processor.js';
+export { txtNodesProcessor } from './txt-nodes-processor.js';
