+++ core/template-processors/ref-processor.js
@@ -1,0 +1,10 @@
+import { DICT } from '../dictionary.js';
+
+/** @type {import('./typedef.js').TplProcessor} */
+export function refProcessor(fr, fnCtx) {
+  [...fr.querySelectorAll(`[${DICT.EL_REF_ATTR}]`)].forEach((/** @type {HTMLElement} */ el) => {
+    let refName = el.getAttribute(DICT.EL_REF_ATTR);
+    fnCtx.ref[refName] = el;
+    el.removeAttribute(DICT.EL_REF_ATTR);
+  });
+}
