+++ core/template-processors/typedef.js
@@ -1,0 +1,5 @@
+/** @typedef {import('../BaseComponent.js').BaseComponent} BaseComponent */
+
+/** @typedef {<T extends BaseComponent = BaseComponent>(fr: DocumentFragment, fnCtx: T) => void} TplProcessor */
+
+export {};
